refactor(api): move route
