+++ api/lib/application/account-recovery/index.js
@@ -37,22 +37,6 @@ const register = async function (server) {
         tags: ['api', 'account-recovery'],
       },
     },
-    {
-      method: 'GET',
-      path: '/api/account-recovery/{temporaryKey}',
-      config: {
-        auth: false,
-        handler: accountRecoveryController.checkAccountRecoveryDemand,
-        validate: {
-          params: Joi.object({ temporaryKey: Joi.string().min(32) }),
-        },
-        notes: [
-          '- Permet de vérifier la demande de récupération de son compte Pix.\n' +
-            '- Renvoie l’utilisateur correspondant à la demande pour une réinitialisation de mot de passe.',
-        ],
-        tags: ['api', 'account-recovery'],
-      },
-    },
   ]);
 };
