+++ api/src/identity-access-management/application/account-recovery/account-recovery.route.js
@@ -32,4 +32,20 @@ export const accountRecoveryRoutes = [
       tags: ['identity-access-management', 'api', 'account-recovery'],
     },
   },
+  {
+    method: 'GET',
+    path: '/api/account-recovery/{temporaryKey}',
+    config: {
+      auth: false,
+      handler: (request, h) => accountRecoveryController.checkAccountRecoveryDemand(request, h),
+      validate: {
+        params: Joi.object({ temporaryKey: Joi.string().min(32) }),
+      },
+      notes: [
+        '- Permet de vérifier la demande de récupération de son compte Pix.\n' +
+          '- Renvoie l’utilisateur correspondant à la demande pour une réinitialisation de mot de passe.',
+      ],
+      tags: ['identity-access-management', 'api', 'account-recovery'],
+    },
+  },
 ];
