+++ api/tests/identity-access-management/acceptance/application/account-recovery/account-recovery.route.test.js
@@ -1,11 +1,16 @@
 import { createServer, databaseBuilder, expect, knex } from '../../../../test-helper.js';
 
 describe('Acceptance | Identity Access Management | Application | Route | account-recovery', function () {
+  let server;
+
+  beforeEach(async function () {
+    server = await createServer();
+  });
+
   describe('PATCH /api/account-recovery', function () {
     context('when user has pix authentication method', function () {
       it("proceeds to the account recover by changing user's password and email", async function () {
         // given
-        const server = await createServer();
         const userId = databaseBuilder.factory.buildUser.withRawPassword({
           email: 'old-email@example.net',
           rawPassword: 'oldPassword',
@@ -52,7 +57,6 @@ describe('Acceptance | Identity Access Management | Application | Route | accoun
     context('when user has no pix authentication method', function () {
       it('proceeds to the account recover by create pix authentication method', async function () {
         // given
-        const server = await createServer();
         const userWithGarAuthenticationMethod = databaseBuilder.factory.buildUser.withoutPixAuthenticationMethod({
           cgu: false,
         });
@@ -95,6 +99,42 @@ describe('Acceptance | Identity Access Management | Application | Route | accoun
         expect(newUserEmail).to.equal('new-email@example.net');
         expect(cgu).to.equal(true);
         expect(userAuthenticationMethods).to.deepEqualArray([{ identityProvider: 'GAR' }, { identityProvider: 'PIX' }]);
+      });
+    });
+  });
+
+  describe('GET /api/account-recovery/{temporaryKey}', function () {
+    context('when account recovery demand found', function () {
+      it('returns 200 http status code', async function () {
+        // given
+        const temporaryKey = 'FfgpFXgyuO062nPUPwcb8Wy3KcgkqR2p2GyEuGVaNI4=';
+        const userId = 1234;
+        const newEmail = 'newEmail@example.net';
+        const firstName = 'Gertrude';
+        databaseBuilder.factory.buildUser.withRawPassword({ id: userId });
+        const { id: organizationLearnerId } = databaseBuilder.factory.buildOrganizationLearner({ userId, firstName });
+        const { id } = databaseBuilder.factory.buildAccountRecoveryDemand({
+          userId,
+          organizationLearnerId,
+          temporaryKey,
+          newEmail,
+          used: false,
+        });
+        await databaseBuilder.commit();
+
+        const options = {
+          method: 'GET',
+          url: '/api/account-recovery/' + temporaryKey,
+        };
+
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result.data.attributes.email).to.equal(newEmail.toLowerCase());
+        expect(response.result.data.attributes['first-name']).to.equal(firstName);
+        expect(response.result.data.id).to.equal(id.toString());
       });
     });
   });
